Add files via upload
n immediate-precedence, single-stage model of parallel units is described. The model is inspired by an immediate precedence MILP model developed by Kopanos et al., (19) for the integrated planning and scheduling problem of parallel continuous processes. Instead of using the idea of the mixed discrete-continuous time representation, the proposed MILP model relies on a unified time horizon. Furthermore, efficient big-M values have been investigated in order to improve the overall computational performance of the model. Except for the typical assignment, timing and sequencing constraints and problem-specific constraints have also been included. Henceforth, we will also refer to this immediate precedence MILP model as IPM. A detailed description of the MILP model is presented below, as follows:
Assignment constraints
	(3.1.1)
Constraints 3.1.1 guarantee that each product order is assigned to one unit j ∈ Ji.
Product orders sequencing constraints
	(3.1.2)
	(3.1.3)
	(3.1.4)
We introduce the binary variable XXi′,i,j to define the local immediate precedence between two products i and i′. The binary variable takes the value 1, only if product order i′ is processed immediately after production order i in unit j ∈ J. Constraints 3.1.2 and 3.1.3 ensure that if production order i ∈ I is allocated to packing line j ∈ J, at most one production order is processed before and after it, respectively. Apparently, in case that the production order is processed first or last then it has no predecessor or successor. According to constraint 3.1.4, the total number of sequences in a packing unit j ∈ J has to be equal to the total number of produced orders minus one.
Time window constraints
	(3.1.5)
According to the underlying inventory constraints of the plant, some products have to be produced during a strictly defined time window. These production campaign cannot start before a lower time limit, Loweri, without also exceeding their related due dates DDATEi. The parameter windowi takes the value 1, only if a production order has to be produced during a specific time window as the value Loweri is predefined according to the industry’s needs. Otherwise, the windowi parameter takes the value 0 and the parameter Loweri is also equal to 0.
Timing constraints
	(3.1.6)
According to the big-M constraints 3.1.6, the completion time Ci′ of product order i′ has to be greater than the completion time of whichever product i is produced beforehand at the same unit plus the processing time Ti′ and the corresponding changeover time, expressed by the parameter changeoveri,i′, only if the binary variable XXi,i′,j is equal to 1. Constraint 3.1.6 also ensures the avoidance of sequence subcycles in the final schedule.
	(3.1.7)
Constraint 3.1.7 enforces the completion time Ci′ of product order i′ to be smaller or equal to the sum of the completion time Ci of product i produced prior to order i′ at the same unit, the processing time Ti′, and the changeover time, changeoveri,i′. Constraint 3.1.7 has to be taken into account if the minimization of changeover times constitutes the objective function because otherwise unnecessary idle times are observed in the generated schedules. The production orders produced in a strict time window (windowi = 1) are excluded from constraints 3.1.7 because infeasible production schedules may be generated.
The selection of the big-M value has a crucial impact on the computational complexity. (16) Increased big-M values relax the domain of the continuous variables. The same pattern appears also in the solution of the MILP problem. Hence, the computational time required for achieving the global optimum and reducing the relative gap between the relaxed and the best integer solution is getting prohibitively high. (37) In the proposed MILP formulation efficient big-M values have been chosen for the purpose of reducing the computational time and providing more efficient mathematical models.
In particular, according to constraints 3.1.6, the big-M value is equal to the DDATEi. Because variable Ci is smaller than or equal to the parameter DDATEi and the binary variable XXi,i′,j is equal to 0, the constraint expresses that Ci′ should be greater than the related processing time, Ti′, minus a small number (equal to Ci – DDATEi), which is significantly smaller than the usually used value of the time horizon under consideration.
The same concept is also applied in constraint 3.1.7, as the big-M value is equal to (DDATEi′ – Ti′). When the binary variable XXi,i′,j is equal to 0, the constraint expresses that the Ci′ should be smaller than or equal to the related due date time (DDATEi′) plus the value of the variable Ci. The proposed value is also significant smaller than the commonly utilized value and equal to the scheduling time horizon of interest.
Formula card constraints (production stage constraints)
	(3.1.8)
Constraint 3.1.8 is referred to the production/formulation stage of the facility. As described above, several changeovers take place among the production of different intermediate products due to cleaning or other activities. Furthermore, a significant amount of the byproduct waste material is generated, which can be partially reused into the next product campaigns. Because of the lack of necessary resources and the limited storage capacity of the occurred waste, the total changeovers related to the intermediate products with a different recipe, formulai, should not exceed an upper limit, Limit.
Due date constraints
	(3.1.9)
Constraint 3.1.9 forces the completion time of production order Ci to be lower than or equal to its deadline, expressed by the parameter DDATEi.
Objective functions
(a) Minimization of production makespan
	(3.1.10)
(b) Minimization of total changeover time
	(3.1.11)
Objective 3.1.10 expresses the minimization of the total production makespan, Cmax, while constraint 3.1.11 expresses the minimization of changeovers and unnecessary idle times.
Plant’s maintenance activities are also considered. Each maintenance task is represented by a dummy product order, which is inserted into the production schedule, by fixing their allocation and their completion variables Yi,j and Ci. These dummy product orders also need to be processed during a time window, and therefore their related parameter windowi is equal to 1.